# improve symbol table performance