@Spacechild1 2020-10-01
Also, we could start with a larger table size, like 4096 or 8192, to avoid the first couple of rehashings. As I've mentioned, Pd already *starts* with more than 500 symbols. What do you think?