@millerpuckette 2020-10-01
I've been thinking abut this for years... I think 500 symbols out of about 1500 (which would start to load the table down) is a reasonable fraction but at the same time I think 124 is a bit small for these days.  At the same time I try to keep Pd's footprint down to something that will fit in an Arduino-sized system.  On the whole I'd support a "reasonable" increase in default size and/or make the default symtab size easily reconfigurable at compile time.

I wouldn't want ever to have Pd stop to rehash - that creates a CPU spike.  OTOH one could leave the symtab intact, but once it exceeds a certain fill level just start another bigger one, ad nauseam.

I'd put the  symbol table stuff in a local struct to keep things out of the pd instance structure.