@Spacechild1 2020-10-01
> On the whole I'd support a "reasonable" increase in default size and/or make the default symtab size easily reconfigurable at compile time.

Sounds good!

> OTOH one could leave the symtab intact, but once it exceeds a certain fill level just start another bigger one, ad nauseam.

That's not a bad idea, actually... With a more aggressive grow factor this could work (otherwise we still end up with long linked lists). However, there will always be *some* point, where allocating a new symbol table (without rehashing) will cause a CPU spike (a table of 100 000 elements certainly will), but at least it will happen only in extreme cases.

I can make another PR and benchmark against this one.