@umlaeute 2020-10-02
> I wouldn't want ever to have Pd stop to rehash - that creates a CPU spike.

the problem (obviously) comes from the possibility, that we run out of run free slots during runtime.
this is similar to resizing an array (e.g. because you load a larger soundfile) during runtime.
the solution for the latter is to simply make sure (at init time) that the arrays are big enough, so they don't need resizing during runtime.

could we add something like that for the symboltable as well?

~~~
[; pd symboltablesize 65536(
~~~