@Ant1r 2020-10-02
(should we continue to discuss here? #1172 seems having only advantages over this one)

I agree we can hardly take the risk of rehashing while audio is being performed.

We also could manually ask Pd to rehash if needed, like:
`[; pd rehash(`
I was thinking that rehash-if-needed could be automatically triggered each time DSP is shut off.

Also: the initial symbol table size could also be settable via a new startup flag (e.g` -symboltablesize 65536`); and personally I would be fine with a default initial size of 65536...

Whatever: I think (as a patch writer) it would be great if we could query the current number of symbols, for debugging purpose, e.g with [pdcontrol].