@umlaeute 2020-10-02
> We also could manually ask Pd to rehash if needed, like:
> `[; pd rehash(`
> I was thinking that rehash-if-needed could be automatically triggered each time DSP is shut off.

ah well.
but if you know that you are going to need 100000 symbols, that doesn't really help you much.
the symbol generation will become slower and slower over time, and once it is slow enough you must manually trigger a re-hash so that it gets fast again.
if you can specify the hashsize beforehand, you would never get slow in the fist place.