@Spacechild1 2020-10-02
Please see https://github.com/pure-data/pure-data/pull/1172 for a solution that doesn't require rehashing. The symbol table expands seemlessly, so there shouldn't ever be a need to manually specifiy the symbol table size, let alone do a manual rehash!

I think that the symbol table is an implementation detail that shouldn't be exposed in the public API. It should work well in all cases without requiring the user to know how it works internally.

> I was thinking that rehash-if-needed could be automatically triggered each time DSP is shut off.

Actually, that brings me to an idea for my other implementation.

I'm closing this PR in favor of https://github.com/pure-data/pure-data/pull/1172