@Spacechild1 2020-10-02
> Actually, that brings me to an idea for my other implementation

Done.